+++ src/EditorPane/ShotFrame.jsx
