+++ src/EditorPane/Snippet.js
