+++ src/EditorPane/codemirror-hint-extension.js
