+++ src/EditorPane/index.js
@@ -1,0 +1,2 @@
+export { default } from './EditorPane';
+export { MimeTypes } from './Editor';
