+++ src/Monitor/PaletteDialog.jsx
