+++ src/Monitor/index.js
@@ -1,0 +1,3 @@
+export { default } from './Monitor';
+export { default as Sizer } from './Sizer';
+export { default as Menu } from './Menu';
